Add function init_kernels
Add general level function init_kernels() to compute initial kernels
with k-means++, random, or a function given as an argument. It is also
possible to recieve an array as the initial kernels; in that case, the
function just make sure that the given array has the same dtype as the
training data.