Refactor the crossvalidation script
Create a new script to run the k-fold crossvalidation of each
hyperparameter point in a single execution. It uses the multiprocessing
module to run each fold in parallel on a child process. The idea here is
to reduce bookkeeping and simplify the configuration and execution of
the crossvalidation.

crossval_config.py is consequently modified, moving all the computation
of fold indices to crossval.py, and removing several old comments.

Since we are really using only the average MCC value during
crossvalidation, only that value for MAP and ML are saved. The
computation of F1, precision, recall, and accuracy would be done only
when training and testing, using the selected hyperparameter point.