Refactor cluster_asignment and likelihood
This fixes the way the likelihood is computed. One should take data from
one class (e.g., preictal), apply CSP-1 (maximizes energy of preictal)
and CSP-C (maximizes energy of interictal), and then cluster the
segments using the corresponding codebooks (C1 for CSP-1 and C2 for
CSP-C).

cluster_asignment also changes so that during test one gets the cluster
assignment for C1 and the test data filtered using CSP-1, and for C2 and
the test data filtered with CSP-C.