Fix training and testing
This fixes the main methodology. For training, data from a known class
is spatially filtered with both CSP-1 (preictal) and CSP-C (interictal),
and their output is clustered with their corresponding codebook (C1 for
CSP-1 and C2 for CSP-C). The likelihood is estimated using those class
conditional cluster assignments.

For testing, the data from both preictal and interictal are concatenated
and passed through both CSP filters, and their output clustered using
the corresponding codebook. The cluster assignment pairs are then
passed to a naive Bayes classifier that estimates the class label of
each segment using the maximum likelihood estimate.