Refactor to fix methodology and to avoid the use of config files
A *.py module calls a `configure_experiment(args)` function defined in a
corresponding config*.py file, avoiding the need of creating a *.yaml
file. Also, dict4cv.py creates the file `entropy.txt`, which saves an
entropy value that is used to create the random generator used in the
*.py modules that need such random generator (e.g., to create the
cross-validation splits).

TODO: We could probably merge some of the configure_experiment()
functions into a single one.